SponsorBlock: Refactored SponsorBlock settings
Also added an experimental setting to disable tablet UI.
