+++ app/src/main/java/org/schabi/newpipe/settings/SponsorBlockCategoriesSettingsFragment.java
@@ -1,12 +1,13 @@
 package org.schabi.newpipe.settings;
 
-import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.widget.Toast;
 
 import androidx.annotation.ColorRes;
 import androidx.annotation.Nullable;
 import androidx.annotation.StringRes;
+import androidx.appcompat.app.AlertDialog;
 import androidx.preference.Preference;
 
 import org.schabi.newpipe.R;
@@ -22,48 +23,52 @@ public class SponsorBlockCategoriesSettingsFragment extends BasePreferenceFragme
     public void onCreatePreferences(final Bundle savedInstanceState, final String rootKey) {
         addPreferencesFromResource(R.xml.sponsor_block_category_settings);
 
-        final Preference resetPreferenceView =
+        final Preference resetPreference =
                 findPreference(getString(R.string.sponsor_block_category_reset_key));
-        if (resetPreferenceView != null) {
-            resetPreferenceView.setOnPreferenceClickListener(preference -> {
-                final Context context = getContext();
+        resetPreference.setOnPreferenceClickListener(p -> {
+            new AlertDialog.Builder(p.getContext())
+                    .setMessage(R.string.sponsor_block_confirm_reset_colors)
+                    .setPositiveButton(R.string.yes, (dialog, which) -> {
+                        final SharedPreferences.Editor editor =
+                                getPreferenceManager()
+                                        .getSharedPreferences()
+                                        .edit();
 
-                if (context != null) {
-                    final SharedPreferences.Editor editor =
-                            getPreferenceManager()
-                                    .getSharedPreferences()
-                                    .edit();
+                        setColorPreference(editor,
+                                R.string.sponsor_block_category_sponsor_color_key,
+                                R.color.sponsor_segment);
 
-                    setColorPreference(editor,
-                            R.string.sponsor_block_category_sponsor_color_key,
-                            R.color.sponsor_segment);
+                        setColorPreference(editor,
+                                R.string.sponsor_block_category_intro_color_key,
+                                R.color.intro_segment);
 
-                    setColorPreference(editor,
-                            R.string.sponsor_block_category_intro_color_key,
-                            R.color.intro_segment);
+                        setColorPreference(editor,
+                                R.string.sponsor_block_category_outro_color_key,
+                                R.color.outro_segment);
 
-                    setColorPreference(editor,
-                            R.string.sponsor_block_category_outro_color_key,
-                            R.color.outro_segment);
+                        setColorPreference(editor,
+                                R.string.sponsor_block_category_interaction_color_key,
+                                R.color.interaction_segment);
 
-                    setColorPreference(editor,
-                            R.string.sponsor_block_category_interaction_color_key,
-                            R.color.interaction_segment);
+                        setColorPreference(editor,
+                                R.string.sponsor_block_category_self_promo_color_key,
+                                R.color.self_promo_segment);
 
-                    setColorPreference(editor,
-                            R.string.sponsor_block_category_self_promo_color_key,
-                            R.color.self_promo_segment);
+                        setColorPreference(editor,
+                                R.string.sponsor_block_category_non_music_color_key,
+                                R.color.non_music_segment);
 
-                    setColorPreference(editor,
-                            R.string.sponsor_block_category_non_music_color_key,
-                            R.color.non_music_segment);
+                        editor.apply();
 
-                    editor.apply();
-                }
-
-                return true;
-            });
-        }
+                        Toast.makeText(p.getContext(), R.string.sponsor_block_reset_colors_toast,
+                                Toast.LENGTH_SHORT).show();
+                    })
+                    .setNegativeButton(R.string.no, (dialog, which) -> {
+                        dialog.dismiss();
+                    })
+                    .show();
+            return true;
+        });
     }
 
     private void setColorPreference(final SharedPreferences.Editor editor,
