+++ app/src/main/java/org/schabi/newpipe/util/DeviceUtils.java
@@ -2,6 +2,7 @@ package org.schabi.newpipe.util;
 
 import android.app.UiModeManager;
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.os.BatteryManager;
@@ -10,8 +11,10 @@ import android.view.KeyEvent;
 
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
+import androidx.preference.PreferenceManager;
 
 import org.schabi.newpipe.App;
+import org.schabi.newpipe.R;
 
 public final class DeviceUtils {
 
@@ -53,6 +56,13 @@ public final class DeviceUtils {
     }
 
     public static boolean isTablet(@NonNull final Context context) {
+        final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
+
+        if (prefs.getBoolean(context
+                .getString(R.string.disable_tablet_ui_key), false)) {
+            return false;
+        }
+
         return (context
                 .getResources()
                 .getConfiguration().screenLayout & Configuration.SCREENLAYOUT_SIZE_MASK)
