+++ app/src/main/res/values/settings_keys.xml
@@ -266,28 +266,6 @@
     <string name="downloads_storage_ask" translatable="false">downloads_storage_ask</string>
     <string name="storage_use_saf" translatable="false">storage_use_saf</string>
 
-    <string name="sponsor_block_home_page_key" translatable="false">sponsor_block_home_page</string>
-    <string name="sponsor_block_enable_key" translatable="false">sponsor_block_enable</string>
-    <string name="sponsor_block_api_url_key" translatable="false">sponsor_block_api_url</string>
-    <string name="sponsor_block_notifications_key" translatable="false">sponsor_block_notifications</string>
-    <string name="sponsor_block_privacy_key" translatable="false">sponsor_block_privacy</string>
-    <string name="sponsor_block_categories_key" translatable="false">sponsor_block_categories</string>
-    <string name="sponsor_block_category_reset_key" translatable="false">sponsor_block_category_reset</string>
-    <string name="sponsor_block_category_sponsor_key" translatable="false">sponsor_block_category_sponsor</string>
-    <string name="sponsor_block_category_sponsor_color_key" translatable="false">sponsor_block_category_sponsor_color</string>
-    <string name="sponsor_block_category_intro_key" translatable="false">sponsor_block_category_intro</string>
-    <string name="sponsor_block_category_intro_color_key" translatable="false">sponsor_block_category_intro_color</string>
-    <string name="sponsor_block_category_outro_key" translatable="false">sponsor_block_category_outro</string>
-    <string name="sponsor_block_category_outro_color_key" translatable="false">sponsor_block_category_outro_color</string>
-    <string name="sponsor_block_category_interaction_key" translatable="false">sponsor_block_category_interaction</string>
-    <string name="sponsor_block_category_interaction_color_key" translatable="false">sponsor_block_category_interaction_color</string>
-    <string name="sponsor_block_category_self_promo_key" translatable="false">sponsor_block_category_self_promo</string>
-    <string name="sponsor_block_category_self_promo_color_key" translatable="false">sponsor_block_category_self_promo_color</string>
-    <string name="sponsor_block_category_non_music_key" translatable="false">sponsor_block_category_music</string>
-    <string name="sponsor_block_category_non_music_color_key" translatable="false">sponsor_block_category_music_color</string>
-    <string name="sponsor_block_whitelist_key" translatable="false">sponsor_block_whitelist</string>
-    <string name="sponsor_block_clear_whitelist_key" translatable="false">sponsor_block_clear_whitelist</string>
-
     <!-- FileName Downloads  -->
     <string name="settings_file_charset_key" translatable="false">file_rename_charset</string>
     <string name="settings_file_replacement_character_key" translatable="false">file_replacement_character</string>
@@ -360,6 +338,32 @@
     <!-- Updates -->
     <string name="update_app_key" translatable="false">update_app_key</string>
     <string name="update_pref_screen_key" translatable="false">update_pref_screen_key</string>
+
+    <!-- SponsorBlock -->
+    <string name="sponsor_block_home_page_key" translatable="false">sponsor_block_home_page</string>
+    <string name="sponsor_block_enable_key" translatable="false">sponsor_block_enable</string>
+    <string name="sponsor_block_api_url_key" translatable="false">sponsor_block_api_url</string>
+    <string name="sponsor_block_notifications_key" translatable="false">sponsor_block_notifications</string>
+    <string name="sponsor_block_privacy_key" translatable="false">sponsor_block_privacy</string>
+    <string name="sponsor_block_categories_key" translatable="false">sponsor_block_categories</string>
+    <string name="sponsor_block_category_reset_key" translatable="false">sponsor_block_category_reset</string>
+    <string name="sponsor_block_category_sponsor_key" translatable="false">sponsor_block_category_sponsor</string>
+    <string name="sponsor_block_category_sponsor_color_key" translatable="false">sponsor_block_category_sponsor_color</string>
+    <string name="sponsor_block_category_intro_key" translatable="false">sponsor_block_category_intro</string>
+    <string name="sponsor_block_category_intro_color_key" translatable="false">sponsor_block_category_intro_color</string>
+    <string name="sponsor_block_category_outro_key" translatable="false">sponsor_block_category_outro</string>
+    <string name="sponsor_block_category_outro_color_key" translatable="false">sponsor_block_category_outro_color</string>
+    <string name="sponsor_block_category_interaction_key" translatable="false">sponsor_block_category_interaction</string>
+    <string name="sponsor_block_category_interaction_color_key" translatable="false">sponsor_block_category_interaction_color</string>
+    <string name="sponsor_block_category_self_promo_key" translatable="false">sponsor_block_category_self_promo</string>
+    <string name="sponsor_block_category_self_promo_color_key" translatable="false">sponsor_block_category_self_promo_color</string>
+    <string name="sponsor_block_category_non_music_key" translatable="false">sponsor_block_category_music</string>
+    <string name="sponsor_block_category_non_music_color_key" translatable="false">sponsor_block_category_music_color</string>
+    <string name="sponsor_block_whitelist_key" translatable="false">sponsor_block_whitelist</string>
+    <string name="sponsor_block_clear_whitelist_key" translatable="false">sponsor_block_clear_whitelist</string>
+
+    <!-- Extras -->
+    <string name="disable_tablet_ui_key" translatable="false">disable_tablet_ui</string>
 
     <!-- Localizations   -->
     <string name="default_localization_key" translatable="false">system</string>
