+++ app/src/release/res/xml/main_settings.xml
@@ -46,4 +46,16 @@
         android:key="update_pref_screen_key"
         android:title="@string/settings_category_updates_title"
         app:iconSpaceReserved="false" />
+
+    <PreferenceScreen
+        android:fragment="org.schabi.newpipe.settings.SponsorBlockSettingsFragment"
+        android:icon="?attr/ic_sponsor_block_enable"
+        android:title="@string/sponsor_block"
+        app:iconSpaceReserved="false" />
+
+    <PreferenceScreen
+        android:fragment="org.schabi.newpipe.settings.ExtraSettingsFragment"
+        android:icon="?attr/ic_asterisk"
+        android:title="@string/extras"
+        app:iconSpaceReserved="false" />
 </PreferenceScreen>
